Fixed the coordinates to be doubles (because that is what the WII remote returns)

+++ RXTXRobot/src/rxtxrobot/Coord.java
@@ -2,36 +2,36 @@ package rxtxrobot;
 
 public class Coord
 {
-    private int x;
-    private int y;
-    private int z;
-    public Coord(int x, int y, int z)
+    private double x;
+    private double y;
+    private double z;
+    public Coord(double x, double y, double z)
     {
         this.x = x;
         this.y = y;
         this.z = z;
     }
-    public int getX()
+    public double getX()
     {
         return x;        
     }
-    public int getY()
+    public double getY()
     {
         return y;
     }
-    public int getZ()
+    public double getZ()
     {
         return z;
     }
-    public void setX(int x)
+    public void setX(double x)
     {
         this.x = x;
     }
-    public void setY(int y)
+    public void setY(double y)
     {
         this.y = y;
     }
-    public void setZ(int z)
+    public void setZ(double z)
     {
         this.z = z;
     }
